refactor: suppply use cases
