--- app/src/main/java/com/bulletapps/candypricer/domain/usecase/DeleteSupplyUseCaseImpl.kt
@@ -1,10 +1,0 @@
-package com.bulletapps.candypricer.domain.usecase
-
-import com.bulletapps.candypricer.domain.model.Supply
-import com.bulletapps.candypricer.data.repository.CandyPricerRepository
-
-class DeleteSupplyUseCaseImpl(
-    private val repository: CandyPricerRepository
-) : DeleteSupplyUseCase {
-    override suspend fun invoke(supply: Supply) = repository.deleteSupply(supply)
-}
