--- app/src/main/java/com/bulletapps/candypricer/domain/usecase/GetAllSuppliesUseCase.kt
@@ -1,7 +1,0 @@
-package com.bulletapps.candypricer.domain.usecase
-
-import com.bulletapps.candypricer.domain.model.Supply
-
-interface GetAllSuppliesUseCase {
-    suspend operator fun invoke() : List<Supply>
-}
