--- app/src/main/java/com/bulletapps/candypricer/domain/usecase/GetSupplyUseCase.kt
@@ -1,7 +1,0 @@
-package com.bulletapps.candypricer.domain.usecase
-
-import com.bulletapps.candypricer.domain.model.Supply
-
-interface GetSupplyUseCase {
-    suspend operator fun invoke(id: Int) : Supply?
-}
