--- app/src/main/java/com/bulletapps/candypricer/domain/usecase/UpdateSupplyUseCase.kt
@@ -1,7 +1,0 @@
-package com.bulletapps.candypricer.domain.usecase
-
-import com.bulletapps.candypricer.domain.model.Supply
-
-interface UpdateSupplyUseCase {
-    suspend operator fun invoke (supply: Supply)
-}
