--- app/src/main/java/com/bulletapps/candypricer/domain/usecase/UpdateSupplyUseCaseImpl.kt
@@ -1,11 +1,0 @@
-package com.bulletapps.candypricer.domain.usecase
-
-import com.bulletapps.candypricer.domain.model.Supply
-import com.bulletapps.candypricer.data.repository.CandyPricerRepository
-import javax.inject.Inject
-
-class UpdateSupplyUseCaseImpl @Inject constructor(
-    private val repository: CandyPricerRepository
-) : UpdateSupplyUseCase {
-    override suspend fun invoke(supply: Supply) = repository.updateSupply(supply)
-}
