+++ app/src/main/java/com/bulletapps/candypricer/domain/usecase/supply/CreateSupplyUseCase.kt
@@ -1,12 +1,11 @@
-package com.bulletapps.candypricer.domain.usecase
+package com.bulletapps.candypricer.domain.usecase.supply
 
 import com.bulletapps.candypricer.data.parameters.CreateSupplyParameters
-import com.bulletapps.candypricer.domain.model.Supply
 import com.bulletapps.candypricer.data.repository.CandyPricerRepository
 import javax.inject.Inject
 
 class CreateSupplyUseCase @Inject constructor(
     private val repository: CandyPricerRepository
 ) {
-    suspend fun invoke(supply: CreateSupplyParameters) = repository.createSupply(supply)
+    suspend operator fun invoke(supply: CreateSupplyParameters) = repository.createSupply(supply)
 }
