+++ app/src/main/java/com/bulletapps/candypricer/presentation/ui/scenes/main/user/login/ScreenLogin.kt
@@ -96,7 +96,7 @@ private fun Screen(
 @Composable
 private fun DisplayToast(uiState: UIState) {
     val toastMessage by uiState.textToast.collectAsState()
-    Toast(toastMessage)
+    Toast(toastMessage.asString())
 }
 
 @Composable
